bmi_closed_loop/requirements.txt
flask
psycopg2-binary
valkey
openpyxl
